requirements.txt
# object detection mqtt

paho-mqtt               # publish notification of object detection
opencv-python           # cv2
imutils
numpy
